flask_digital_analysis/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.0
beautifulsoup4==4.12.2
certifi==2023.11.17
python-multipart==0.0.6
python-dotenv==1.0.0
lxml==5.3.0
anthropic>=0.34.0
instaloader>=4.9.5
pandas>=1.5.0
requests
selenium>=4.15.0
textblob>=0.17.1
plotly>=5.17.0
transformers>=4.41.0
colorthief>=0.2.1
Pillow>=9.0.0
gunicorn
Flask==3.0.0
streamlit
webdriver-manager>=4.0.0
torch>=2.2.0